Basic population movement

+++ API/AwdagAPI/Domain/Models/Entities/Food.cs
@@ -3,8 +3,8 @@
     public class Food
     {
         public int Id { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
         public virtual Aquarium Aquarium { get; set; }
         public int AquariumId { get; set; }
     }
